04_10_2021
Note: Currently, the advanced encryption standard (AES) is being utilized in lieu of the earlier data encryption standard (DES). Implementation is still taking place within the first phase of the PKI-based authentication protocol as thorough testing of any phase has yet to occur. That is, the current state of the program is far from finished and certainly unusable.