04_10_2021
Note: The first phase of the protocol, which involves the registration of the application server, is complete; now, I'm working on the second phase of the protocol - parsing the received message contents of the second step within the client registration phase (fourth overall step of protocol). I expect to have completed the second and third phases of the protocol (client registration and service request) by the end of tomorrow, thus leaving only testing and refinement of documentation. It should be noted, however, that all completed steps up to this point have been tested.