04_11_2021
Note: The program will need to be reworked from its current state of development in order to mitigate inconsistencies caused by utilizing AES instead of DES; there is an inherent limitation with the current AES implementation, which has been discovered during the sixth step of the protocol process, that prevents proper communication given that the client is not allocated a private key via the centralized certificate authority. Currently, AES is utilizing RSA keys, which not only tethers these two encryption mechanisms but provides difficulty in the reworking of the current system.