04_13_2021
Note: Code has been restructured to exclude AES implementations; DES is now used as specified in the assignment specifications. Naming conventions have been revised, however, there exists a need to induce complete uniformity across source files. The first and second phases - application server and client registration - are implemented and need only to be tested. The third phase has yet to be implemented. A JPG file has been included in this commitment that delineates the various phases in prescribed steps; this message flow diagram can be useful in understanding, with more granularity, the nature of the project.