04_14_2021
Note: Testing of phase 1 (application server registration) is completed; this portion of the project is functionally finished but in need of a revision of naming-convention uniformity. Hardcoded index boundaries for parsing messages could be discarded for a more dynamic and optimal mechanism, as the current parsing maneuvers are contingent upon size-oriented assumptions.