04_15_2021
Note: This repository commit serves as the final version in the project's developmental iterations; there will be no further updates. One drawback of the implementation is that private keys are stored in globally accessible files. For this reason, the application server and client simply open these respective files to import desired keys. This configuration is partly due to the fact that the pycryptodome library seems to have issues dealing with the importation of keys outside of the generating program. If this project is to be further pursued, it would be suggested that the centralized certificate authority, which generates and stores all RSA keys, restrict access to files containing private keys. Another circumvention would involve the sending of files containing private keys, as determining access to a given private key is already approved for an entity with or without this addition. That is, we can either send text, which the receiving entity will have to parse, or much more simply, we can just send the file containing the private key that has been created for that particular entity.